make new reader based on the previously updated reader (untested)
