--- GCRCatalogs/catalog_configs/dc2_static_coadd.yaml
@@ -1,2 +1,0 @@
-subclass_name: dc2_static_coadd_reader.DC2StaticCoaddCatalog
-included_by_default: true
